+++ trunk/CoMap/Cluster.cpp
@@ -58,18 +58,18 @@ TreeTemplate<Node>* SumClustering::getTree() const
 	return new TreeTemplate<Node>(root);
 }
 
-vector<unsigned int> SumClustering::getBestPair() throw (Exception)
+vector<size_t> SumClustering::getBestPair() throw (Exception)
 {
-	vector<unsigned int> bestPair(2);
+	vector<size_t> bestPair(2);
 	double distMin = -std::log(0.);
-	for(map<unsigned int, Node *>::iterator i = currentNodes_.begin(); i != currentNodes_.end(); i++)
+	for(map<size_t, Node *>::iterator i = currentNodes_.begin(); i != currentNodes_.end(); i++)
   {
-		unsigned int id = i->first;
-		map<unsigned int, Node *>::iterator j = i;
+		size_t id = i->first;
+		map<size_t, Node *>::iterator j = i;
 		j++;
 		for(; j != currentNodes_.end(); j++)
     {
-			unsigned int jd = j->first;
+			size_t jd = j->first;
 			double dist = matrix_(id, jd);
 			if(dist < distMin)
       {
@@ -83,7 +83,7 @@ vector<unsigned int> SumClustering::getBestPair() throw (Exception)
 	mapping_[bestPair[0]] += mapping_[bestPair[1]];
 	return bestPair;	
 }
-vector<double> SumClustering::computeBranchLengthsForPair(const vector<unsigned int>& pair)
+vector<double> SumClustering::computeBranchLengthsForPair(const vector<size_t>& pair)
 {
 	vector<double> d(2);
 	double dist = matrix_(pair[0], pair[1]) / 2.;
@@ -94,7 +94,7 @@ vector<double> SumClustering::computeBranchLengthsForPair(const vector<unsigned 
 	return d;
 }
 
-double SumClustering::computeDistancesFromPair(const vector<unsigned int>& pair, const vector<double>& branchLengths, unsigned int pos)
+double SumClustering::computeDistancesFromPair(const vector<size_t>& pair, const vector<double>& branchLengths, size_t pos)
 {
 	return distance_->getDistanceForPair(mapping_[pos], mapping_[pair[0]]);
 }
@@ -102,11 +102,11 @@ double SumClustering::computeDistancesFromPair(const vector<unsigned int>& pair,
 void SumClustering::finalStep(int idRoot)
 {
 	NodeTemplate<ClusterInfos>* root = new NodeTemplate<ClusterInfos>(idRoot);
-	map<unsigned int, Node*>::iterator it = currentNodes_.begin();
-	unsigned int i1 = it->first;
+	map<size_t, Node*>::iterator it = currentNodes_.begin();
+	size_t i1 = it->first;
 	Node* n1        = it->second;
 	it++;
-	unsigned int i2 = it->first;
+	size_t i2 = it->first;
 	Node* n2        = it->second;
 	double d = matrix_(i1, i2) / 2;
 	root->addSon(n1);
